+++ src/routes/Widget/containers/WidgetView.scss
